@@ -3,6 +3,7 @@
   "version": "0.3.2-alpha.16",
   "description": "",
   "main": "./dist/cjs/zk.js.cjs",
+  "module": "./dist/zk.js.legacy-esm.js",
   "types": "./dist/zk.js.d.mts",
   "exports": {
     "./package.json": "./package.json",
@@ -57,7 +58,6 @@
       "prettier --write"
     ]
   },
-
   "keywords": [
     "sdk",
     "light",
@@ -88,6 +88,11 @@
   },
   "devDependencies": {
     "@lightprotocol/tsconfig": "workspace:*",
+    "@microsoft/api-extractor": "^7.38.3",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-replace": "^5.0.5",
+    "@rollup/plugin-typescript": "^11.1.5",
     "@types/bn.js": "^5.1.0",
     "@types/chai": "^4.3.0",
     "@types/chai-as-promised": "^7.1.5",
@@ -105,12 +110,13 @@
     "mocha": "^10.2.0",
     "prettier": "^2.6.2",
     "rimraf": "^5.0.5",
+    "rollup": "^4.6.0",
     "seedrandom": "^3.0.5",
     "shelljs": "^0.8.5",
     "ts-mocha": "^10.0.0",
     "ts-node": "^10.9.1",
     "tsup": "^8.0.1",
-    "typescript": "^4.3.5",
+    "typescript": "^5.3.2",
     "typescript-collections": "^1.3.3"
   }
 }
