@@ -14,4 +14,6 @@ export * from "./convertCase";
 export * from "./closeVerifierState";
 export * from "./createRustVerifyingKey";
 
-export * as circuitlibjs from "@lightprotocol/circuit-lib.js";
+import * as circuitlibjs from "@lightprotocol/circuit-lib.js";
+export { circuitlibjs };
+
